lab5 first commit

+++ table.c
@@ -4,6 +4,7 @@
  */
 
 #include <stdio.h>
+#include <assert.h>
 #include "util.h"
 #include "table.h"
 
@@ -45,7 +46,7 @@ TAB_table TAB_empty(void)
 void TAB_enter(TAB_table t, void *key, void *value)
 {int index;
  assert(t && key);
- index = ((unsigned)key) % TABSIZE;
+ index = ((unsigned long)key) % TABSIZE;
  t->table[index] = Binder(key, value,t->table[index], t->top);
  t->top = key;
 }
@@ -54,7 +55,7 @@ void *TAB_look(TAB_table t, void *key)
 {int index;
  binder b;
  assert(t && key);
- index=((unsigned)key) % TABSIZE;
+ index=((unsigned long)key) % TABSIZE;
  for(b=t->table[index]; b; b=b->next)
    if (b->key==key) return b->value;
  return NULL;
@@ -65,7 +66,7 @@ void *TAB_pop(TAB_table t) {
   assert (t);
   k = t->top;
   assert (k);
-  index = ((unsigned)k) % TABSIZE;
+  index = ((unsigned long)k) % TABSIZE;
   b = t->table[index];
   assert(b);
   t->table[index] = b->next;
@@ -75,7 +76,7 @@ void *TAB_pop(TAB_table t) {
 
 void TAB_dump(TAB_table t, void (*show)(void *key, void *value)) {
   void *k = t->top;
-  int index = ((unsigned)k) % TABSIZE;
+  int index = ((unsigned long)k) % TABSIZE;
   binder b = t->table[index];
   if (b==NULL) return;
   t->table[index]=b->next;
